requirements.txt
python-docx==1.1.0
PyYAML==6.0.2
pydantic